refactor: organize vault-mcp into domain subdirectories (#20)
Group 12 flat source files into three domain folders:
- vault-operations/ — vault content CRUD (filesystem, patcher, memory, daily notes)
- search/ — FTS5 indexing + file watcher
- auth/ — OAuth provider, routes, consent page

Server/transport files (server.ts, mcp-router.ts, tool-definitions.ts)
stay at root as the app wiring layer. Tests colocated with their domain.
auth.test.ts moved to src/__tests__/ (tests parent-level src/auth.ts).

All import paths updated. No logic, behavior, or API changes.
418 tests pass, tsc clean, lint clean.
