+++ src/vault-mcp/__tests__/tool-definitions.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, vi } from "vitest"
 import { registerTools, TOOL_NAMES } from "../tool-definitions.js"
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js"
-import type { SearchIndex } from "../search-index.js"
+import type { SearchIndex } from "../search/search-index.js"
 import { logger } from "../../logger.js"
 
 const ALL_TOOL_NAMES = Object.values(TOOL_NAMES)
