+++ src/vault-mcp/auth/__tests__/oauth-provider.test.ts
