+++ src/vault-mcp/auth/consent-page.ts
