+++ src/vault-mcp/server.ts
@@ -4,10 +4,10 @@ import express from "express"
 import type { Request, Response, NextFunction } from "express"
 import { resolve } from "node:path"
 import { fileURLToPath } from "node:url"
-import { createSearchIndex } from "./search-index.js"
-import { startFileWatcher } from "./file-watcher.js"
-import { createOAuthProvider } from "./oauth-provider.js"
-import { createOAuthRoutes } from "./oauth-routes.js"
+import { createSearchIndex } from "./search/search-index.js"
+import { startFileWatcher } from "./search/file-watcher.js"
+import { createOAuthProvider } from "./auth/oauth-provider.js"
+import { createOAuthRoutes } from "./auth/oauth-routes.js"
 import { createMcpRouter } from "./mcp-router.js"
 import { logger } from "../logger.js"
