+++ src/vault-mcp/vault-operations/__tests__/daily-notes.test.ts
@@ -10,7 +10,7 @@ import {
   getDailyNote,
   clearConfigCache,
 } from "../daily-notes.js"
-import { logger } from "../../logger.js"
+import { logger } from "../../../logger.js"
 
 // ── momentToLuxonFormat ──────────────────────────────────────────
